Move basic git methods into separate base class
