+++ clowder/clowder/git/repo.py
@@ -11,7 +11,6 @@ from termcolor import colored, cprint
 
 import clowder.util.formatting as fmt
 from clowder.error.clowder_git_error import ClowderGitError
-from clowder.util.connectivity import is_offline
 from clowder.util.execute import execute_command
 from clowder.util.file_system import remove_directory
 
@@ -59,46 +58,6 @@ class GitRepo(object):
             self._print(' - Abort rebase in progress')
             self._abort_rebase()
 
-    def create_clowder_repo(self, url, branch, depth=0):
-        """Clone clowder git repo from url at path"""
-        if GitRepo.existing_git_repository(self.repo_path):
-            return
-        self._init_repo()
-        self._create_remote(self.remote, url, remove_dir=True)
-        self._checkout_new_repo_branch(branch, depth)
-
-    def configure_remotes(self, upstream_remote_name, upstream_remote_url, fork_remote_name, fork_remote_url):
-        """Configure remotes names for fork and upstream"""
-        if not GitRepo.existing_git_repository(self.repo_path):
-            return
-        try:
-            remotes = self.repo.remotes
-        except GitError:
-            return
-        except (KeyboardInterrupt, SystemExit):
-            sys.exit(1)
-        for remote in remotes:
-            if upstream_remote_url == self._remote_get_url(remote.name):
-                if remote.name != upstream_remote_name:
-                    self._rename_remote(remote.name, upstream_remote_name)
-                    continue
-            if fork_remote_url == self._remote_get_url(remote.name):
-                if remote.name != fork_remote_name:
-                    self._rename_remote(remote.name, fork_remote_name)
-        remote_names = [r.name for r in self.repo.remotes]
-        if upstream_remote_name in remote_names:
-            if upstream_remote_url != self._remote_get_url(upstream_remote_name):
-                actual_url = self._remote_get_url(upstream_remote_name)
-                message = fmt.remote_already_exists_error(upstream_remote_name, upstream_remote_url, actual_url)
-                self._print(message)
-                self._exit(fmt.parallel_exception_error(self.repo_path, message))
-        if fork_remote_name in remote_names:
-            if fork_remote_url != self._remote_get_url(fork_remote_name):
-                actual_url = self._remote_get_url(fork_remote_name)
-                message = fmt.remote_already_exists_error(fork_remote_name, fork_remote_url, actual_url)
-                self._print(message)
-                self._exit(fmt.parallel_exception_error(self.repo_path, message))
-
     def current_branch(self):
         """Return currently checked out branch of project"""
         return self.repo.head.ref.name
@@ -127,12 +86,6 @@ class GitRepo(object):
         """Check if a git submodule exists"""
         return os.path.isfile(os.path.join(path, '.git'))
 
-    @staticmethod
-    def exists(repo_path):
-        """Print existence validation messages"""
-        if not GitRepo.existing_git_repository(repo_path):
-            cprint(' - Project is missing', 'red')
-
     def fetch(self, remote, ref=None, depth=0, remove_dir=False):
         """Fetch from a specific remote ref"""
         remote_output = fmt.remote_string(remote)
@@ -160,122 +113,6 @@ class GitRepo(object):
             self._print(error)
             self._exit(fmt.parallel_exception_error(self.repo_path, error))
         return return_code
-
-    @staticmethod
-    def format_project_ref_string(repo_path):
-        """Return formatted repo ref name"""
-        repo = GitRepo(repo_path, DEFAULT_REMOTE, DEFAULT_REF)
-        local_commits = repo.new_commits()
-        upstream_commits = repo.new_commits(upstream=True)
-        no_local_commits = local_commits == 0 or local_commits == '0'
-        no_upstream_commits = upstream_commits == 0 or upstream_commits == '0'
-        if no_local_commits and no_upstream_commits:
-            status = ''
-        else:
-            local_commits_output = colored('+' + str(local_commits), 'yellow')
-            upstream_commits_output = colored('-' + str(upstream_commits), 'red')
-            status = '[' + local_commits_output + '/' + upstream_commits_output + ']'
-
-        if repo.is_detached():
-            current_ref = repo.sha(short=True)
-            return colored('(HEAD @ ' + current_ref + ')', 'magenta')
-        current_branch = repo.current_branch()
-        return colored('(' + current_branch + ')', 'magenta') + status
-
-    @staticmethod
-    def format_project_string(repo_path, name):
-        """Return formatted project name"""
-        if not GitRepo.existing_git_repository(repo_path):
-            return colored(name, 'green')
-        repo = GitRepo(repo_path, DEFAULT_REMOTE, DEFAULT_REF)
-        if not repo.validate_repo():
-            color = 'red'
-            symbol = '*'
-        else:
-            color = 'green'
-            symbol = ''
-        return colored(name + symbol, color)
-
-    def herd(self, url, depth=0, fetch=True, rebase=False):
-        """Herd ref"""
-        if not GitRepo.existing_git_repository(self.repo_path):
-            self._herd_initial(url, depth=depth)
-            return
-        return_code = self._create_remote(self.remote, url)
-        if return_code != 0:
-            raise ClowderGitError(msg=colored(' - Failed to create remote', 'red'))
-        self._herd(self.remote, self.default_ref, depth=depth, fetch=fetch, rebase=rebase)
-
-    def herd_branch(self, url, branch, depth=0, rebase=False, fork_remote=None):
-        """Herd branch"""
-        if not GitRepo.existing_git_repository(self.repo_path):
-            self._herd_branch_initial(url, branch, depth=depth)
-            return
-        branch_output = fmt.ref_string(branch)
-        branch_ref = 'refs/heads/' + branch
-        if self.existing_local_branch(branch):
-            if self._is_branch_checked_out(branch):
-                self._print(' - Branch ' + branch_output + ' already checked out')
-            else:
-                self._checkout_branch_local(branch)
-            self.fetch(self.remote, depth=depth, ref=branch_ref)
-            if self.existing_remote_branch(branch, self.remote):
-                self._herd_remote_branch(self.remote, branch, depth=depth, rebase=rebase)
-                return
-            if fork_remote:
-                self.fetch(fork_remote, depth=depth, ref=branch_ref)
-                if self.existing_remote_branch(branch, fork_remote):
-                    self._herd_remote_branch(fork_remote, branch, depth=depth, rebase=rebase)
-            return
-        self.fetch(self.remote, depth=depth, ref=branch_ref)
-        if self.existing_remote_branch(branch, self.remote):
-            self._herd(self.remote, branch_ref, depth=depth, fetch=False, rebase=rebase)
-            return
-        else:
-            remote_output = fmt.remote_string(self.remote)
-            self._print(' - No existing remote branch ' + remote_output + ' ' + branch_output)
-        if fork_remote:
-            self.fetch(fork_remote, depth=depth, ref=branch_ref)
-            if self.existing_remote_branch(branch, fork_remote):
-                self._herd(fork_remote, branch_ref, depth=depth, fetch=False, rebase=rebase)
-                return
-            else:
-                remote_output = fmt.remote_string(fork_remote)
-                self._print(' - No existing remote branch ' + remote_output + ' ' + branch_output)
-        fetch = depth != 0
-        self.herd(url, depth=depth, fetch=fetch, rebase=rebase)
-
-    def herd_tag(self, url, tag, depth=0, rebase=False):
-        """Herd tag"""
-        if not GitRepo.existing_git_repository(self.repo_path):
-            self._init_repo()
-            self._create_remote(self.remote, url, remove_dir=True)
-            return_code = self._checkout_new_repo_tag(tag, self.remote, depth)
-            if return_code == 0:
-                return
-            fetch = depth != 0
-            self.herd(url, depth=depth, fetch=fetch, rebase=rebase)
-            return
-        return_code = self.fetch(self.remote, ref='refs/tags/' + tag, depth=depth)
-        if return_code == 0:
-            return_code = self._checkout_tag(tag)
-            if return_code == 0:
-                return
-        fetch = depth != 0
-        self.herd(url, depth=depth, fetch=fetch, rebase=rebase)
-
-    def herd_remote(self, url, remote, branch=None):
-        """Herd remote repo"""
-        return_code = self._create_remote(remote, url)
-        if return_code != 0:
-            raise ClowderGitError(msg=colored(' - Failed to create remote', 'red'))
-        if branch:
-            return_code = self.fetch(remote, ref=branch)
-            if return_code == 0:
-                return
-        return_code = self.fetch(remote, ref=self.default_ref)
-        if return_code != 0:
-            raise ClowderGitError(msg=colored(' - Failed to fetch', 'red'))
 
     def is_detached(self):
         """Check if HEAD is detached"""
@@ -333,55 +170,6 @@ class GitRepo(object):
             print(fmt.command_failed_error(command))
             sys.exit(1)
 
-    def prune_branch_local(self, branch, force):
-        """Prune branch in repository"""
-        branch_output = fmt.ref_string(branch)
-        if branch not in self.repo.heads:
-            self._print(' - Local branch ' + branch_output + " doesn't exist")
-            return
-        prune_branch = self.repo.heads[branch]
-        if self.repo.head.ref == prune_branch:
-            ref_output = fmt.ref_string(GitRepo.truncate_ref(self.default_ref))
-            try:
-                self._print(' - Checkout ref ' + ref_output)
-                self.repo.git.checkout(GitRepo.truncate_ref(self.default_ref))
-            except GitError as err:
-                message = colored(' - Failed to checkout ref', 'red')
-                print(message + ref_output)
-                print(fmt.error(err))
-                sys.exit(1)
-            except (KeyboardInterrupt, SystemExit):
-                sys.exit(1)
-        try:
-            self._print(' - Delete local branch ' + branch_output)
-            self.repo.delete_head(branch, force=force)
-            return
-        except GitError as err:
-            message = colored(' - Failed to delete local branch ', 'red')
-            print(message + branch_output)
-            print(fmt.error(err))
-            sys.exit(1)
-        except (KeyboardInterrupt, SystemExit):
-            sys.exit(1)
-
-    def prune_branch_remote(self, branch, remote):
-        """Prune remote branch in repository"""
-        origin = self._remote(remote)
-        branch_output = fmt.ref_string(branch)
-        if branch not in origin.refs:
-            self._print(' - Remote branch ' + branch_output + " doesn't exist")
-            return
-        try:
-            self._print(' - Delete remote branch ' + branch_output)
-            self.repo.git.push(remote, '--delete', branch)
-        except GitError as err:
-            message = colored(' - Failed to delete remote branch ', 'red')
-            print(message + branch_output)
-            print(fmt.error(err))
-            sys.exit(1)
-        except (KeyboardInterrupt, SystemExit):
-            sys.exit(1)
-
     @staticmethod
     def ref_type(ref):
         """Return branch, tag, sha, or unknown ref type"""
@@ -395,69 +183,11 @@ class GitRepo(object):
             return 'sha'
         return 'unknown'
 
-    def reset(self, depth=0):
-        """Reset branch to upstream or checkout tag/sha as detached HEAD"""
-        if GitRepo.ref_type(self.default_ref) == 'branch':
-            branch = GitRepo.truncate_ref(self.default_ref)
-            branch_output = fmt.ref_string(branch)
-            if not self.existing_local_branch(branch):
-                return_code = self._create_branch_local_tracking(branch, self.remote, depth=depth, fetch=True)
-                if return_code != 0:
-                    message = colored(' - Failed to create tracking branch ', 'red') + branch_output
-                    self._print(message)
-                    self._exit(fmt.parallel_exception_error(self.repo_path, message))
-                return
-            elif self._is_branch_checked_out(branch):
-                self._print(' - Branch ' + branch_output + ' already checked out')
-            else:
-                self._checkout_branch_local(branch)
-            remote_output = fmt.remote_string(self.remote)
-            if not self.existing_remote_branch(branch, self.remote):
-                message = colored(' - No existing remote branch ', 'red') + remote_output + ' ' + branch_output
-                self._print(message)
-                self._exit(fmt.parallel_exception_error(self.repo_path, message))
-            self.fetch(self.remote, ref=self.default_ref, depth=depth)
-            self._print(' - Reset branch ' + branch_output + ' to ' + remote_output + ' ' + branch_output)
-            remote_branch = self.remote + '/' + branch
-            self._reset_head(branch=remote_branch)
-        elif GitRepo.ref_type(self.default_ref) == 'tag':
-            self.fetch(self.remote, ref=self.default_ref, depth=depth)
-            self._checkout_tag(GitRepo.truncate_ref(self.default_ref))
-        elif GitRepo.ref_type(self.default_ref) == 'sha':
-            self.fetch(self.remote, ref=self.default_ref, depth=depth)
-            self._checkout_sha(self.default_ref)
-
     def sha(self, short=False):
         """Return sha for currently checked out commit"""
         if short:
             return self.repo.git.rev_parse(self.repo.head.commit.hexsha, short=True)
         return self.repo.head.commit.hexsha
-
-    def start(self, remote, branch, depth, tracking):
-        """Start new branch in repository"""
-        if branch not in self.repo.heads:
-            if not is_offline():
-                return_code = self.fetch(remote, ref=branch, depth=depth)
-                if return_code != 0:
-                    sys.exit(1)
-            return_code = self._create_branch_local(branch)
-            if return_code != 0:
-                sys.exit(1)
-            return_code = self._checkout_branch_local(branch)
-            if return_code != 0:
-                sys.exit(1)
-        else:
-            branch_output = fmt.ref_string(branch)
-            print(' - ' + branch_output + ' already exists')
-            correct_branch = self._is_branch_checked_out(branch)
-            if correct_branch:
-                print(' - On correct branch')
-            else:
-                return_code = self._checkout_branch_local(branch)
-                if return_code != 0:
-                    sys.exit(1)
-        if tracking and not is_offline():
-            self._create_branch_remote_tracking(branch, remote, depth)
 
     def stash(self):
         """Stash current changes in repository"""
@@ -481,29 +211,6 @@ class GitRepo(object):
             cprint(' - Failed to print status', 'red')
             print(fmt.command_failed_error(command))
             sys.exit(return_code)
-
-    def sync(self, fork_remote, rebase=False):
-        """Sync fork with upstream remote"""
-        if self.print_output:
-            print(' - Sync fork with upstream remote')
-        if GitRepo.ref_type(self.default_ref) != 'branch':
-            message = colored(' - Can only sync branches', 'red')
-            self._print(message)
-            self._exit(fmt.parallel_exception_error(self.repo_path, message))
-        fork_remote_output = fmt.remote_string(fork_remote)
-        branch_output = fmt.ref_string(GitRepo.truncate_ref(self.default_ref))
-        if rebase:
-            self._rebase_remote_branch(self.remote, GitRepo.truncate_ref(self.default_ref))
-        else:
-            self._pull(self.remote, GitRepo.truncate_ref(self.default_ref))
-        self._print(' - Push to ' + fork_remote_output + ' ' + branch_output)
-        command = ['git', 'push', fork_remote, GitRepo.truncate_ref(self.default_ref)]
-        return_code = execute_command(command, self.repo_path, print_output=self.print_output)
-        if return_code != 0:
-            message = colored(' - Failed to push to ', 'red')
-            self._print(message + fork_remote_output + ' ' + branch_output)
-            self._print(fmt.command_failed_error(command))
-            self._exit(fmt.parallel_exception_error(self.repo_path, message, fork_remote_output, ' ', branch_output))
 
     @staticmethod
     def truncate_ref(ref):
@@ -575,9 +282,9 @@ class GitRepo(object):
         self.fetch(self.remote, depth=depth, ref=branch, remove_dir=True)
         if not self.existing_remote_branch(branch, self.remote):
             remove_directory(self.repo_path)
-            message = colored(' - No existing remote branch ', 'red')
-            self._print(message + remote_output + ' ' + branch_output)
-            self._exit(fmt.parallel_exception_error(self.repo_path, message, remote_output, ' ', branch_output))
+            message = colored(' - No existing remote branch ', 'red') + remote_output + ' ' + branch_output
+            self._print(message)
+            self._exit(fmt.parallel_exception_error(self.repo_path, message))
         self._create_branch_local_tracking(branch, self.remote, depth=depth, fetch=False, remove_dir=True)
 
     def _checkout_new_repo_commit(self, commit, remote, depth):
@@ -796,72 +503,8 @@ class GitRepo(object):
     def _exit(self, message):
         """Print output if print_output is True"""
         if self.parallel:
-            sys.exit(1)
-        raise ClowderGitError(msg=message)
-
-    def _herd(self, remote, ref, depth=0, fetch=True, rebase=False):
-        """Herd ref"""
-        if GitRepo.ref_type(ref) == 'branch':
-            branch = GitRepo.truncate_ref(ref)
-            branch_output = fmt.ref_string(branch)
-            if not self.existing_local_branch(branch):
-                return_code = self._create_branch_local_tracking(branch, remote, depth=depth, fetch=fetch)
-                if return_code != 0:
-                    message = colored(' - Failed to create tracking branch ', 'red')
-                    self._exit(fmt.parallel_exception_error(self.repo_path, message, branch_output))
-                return
-            elif self._is_branch_checked_out(branch):
-                self._print(' - Branch ' + branch_output + ' already checked out')
-            else:
-                self._checkout_branch_local(branch)
-            if not self.existing_remote_branch(branch, remote):
-                return
-            if not self._is_tracking_branch(branch):
-                self._set_tracking_branch_commit(branch, remote, depth)
-                return
-            if rebase:
-                self._rebase_remote_branch(remote, branch)
-                return
-            self._pull(remote, branch)
-        elif GitRepo.ref_type(ref) == 'tag':
-            self.fetch(remote, depth=depth, ref=ref)
-            self._checkout_tag(GitRepo.truncate_ref(ref))
-        elif GitRepo.ref_type(ref) == 'sha':
-            self.fetch(remote, depth=depth, ref=ref)
-            self._checkout_sha(ref)
-
-    def _herd_initial(self, url, depth=0):
-        """Herd ref initial"""
-        self._init_repo()
-        self._create_remote(self.remote, url, remove_dir=True)
-        if GitRepo.ref_type(self.default_ref) == 'branch':
-            self._checkout_new_repo_branch(GitRepo.truncate_ref(self.default_ref), depth)
-        elif GitRepo.ref_type(self.default_ref) == 'tag':
-            self._checkout_new_repo_tag(GitRepo.truncate_ref(self.default_ref), self.remote, depth, remove_dir=True)
-        elif GitRepo.ref_type(self.default_ref) == 'sha':
-            self._checkout_new_repo_commit(self.default_ref, self.remote, depth)
-
-    def _herd_branch_initial(self, url, branch, depth=0):
-        """Herd branch initial"""
-        self._init_repo()
-        self._create_remote(self.remote, url, remove_dir=True)
-        self.fetch(self.remote, depth=depth, ref=branch)
-        if not self.existing_remote_branch(branch, self.remote):
-            remote_output = fmt.remote_string(self.remote)
-            self._print(' - No existing remote branch ' + remote_output + ' ' + fmt.ref_string(branch))
-            self._herd_initial(url, depth=depth)
-            return
-        self._create_branch_local_tracking(branch, self.remote, depth=depth, fetch=False, remove_dir=True)
-
-    def _herd_remote_branch(self, remote, branch, depth=0, rebase=False):
-        """Herd remote branch"""
-        if not self._is_tracking_branch(branch):
-            self._set_tracking_branch_commit(branch, remote, depth)
-            return
-        if rebase:
-            self._rebase_remote_branch(remote, branch)
-            return
-        self._pull(remote, branch)
+            raise ClowderGitError(msg=message)
+        sys.exit(1)
 
     def _init_repo(self):
         """Initialize repository"""
@@ -978,7 +621,7 @@ class GitRepo(object):
             sys.exit(1)
 
     def _remote_get_url(self, remote):
-        """Reset head of repo, discarding changes"""
+        """Get url of remote"""
         return self.repo.git.remote('get-url', remote)
 
     def _rename_remote(self, remote_from, remote_to):
@@ -1060,32 +703,6 @@ class GitRepo(object):
             if remove_dir:
                 remove_directory(self.repo_path)
             sys.exit(1)
-
-    def _set_tracking_branch_commit(self, branch, remote, depth):
-        """Set tracking relationship between local and remote branch if on same commit"""
-        branch_output = fmt.ref_string(branch)
-        origin = self._remote(remote)
-        return_code = self.fetch(remote, depth=depth, ref=branch)
-        if return_code != 0:
-            raise ClowderGitError(msg=colored(' - Failed to fech', 'red'))
-        if not self.existing_local_branch(branch):
-            message = colored(' - No local branch ', 'red')
-            self._print(message + branch_output + '\n')
-            self._exit(fmt.parallel_exception_error(self.repo_path, message, branch_output, '\n'))
-        if not self.existing_remote_branch(branch, remote):
-            message = colored(' - No remote branch ', 'red')
-            self._print(message + branch_output + '\n')
-            self._exit(fmt.parallel_exception_error(self.repo_path, message, branch_output, '\n'))
-        local_branch = self.repo.heads[branch]
-        remote_branch = origin.refs[branch]
-        if local_branch.commit != remote_branch.commit:
-            message_1 = colored(' - Existing remote branch ', 'red')
-            message_2 = colored(' on different commit', 'red')
-            self._print(message_1 + branch_output + message_2 + '\n')
-            self._exit(fmt.parallel_exception_error(self.repo_path, message_1, branch_output, message_2, '\n'))
-        return_code = self._set_tracking_branch(remote, branch)
-        if return_code != 0:
-            self._exit(colored(' - Failed to set tracking branch', 'red'))
 
     def _untracked_files(self):
         """Execute command and display continuous output"""
